Remove create-footer logic and add content to html
Like the header, the footer doesn't change with the page content, so
there's no reason to re-render it each time the page is build with a
new JS call (and duplicate the code to do so). So, the logic to render
the footer has been removed from the JS and and placed in index.html.